Moved molecular-specific fields into a separate mixin
